+++ scripts/utility/diceSoNice.js
@@ -1,0 +1,10 @@
+async function early(workflow) {
+    workflow.workflowOptions.damageRollDSN = false
+}
+async function late(workflow) {
+    MidiQOL.displayDSNForRoll(workflow.damageRoll, 'damageRoll')
+}
+export let diceSoNice = {
+    early: early,
+    late: late
+}